chore(analysis-cached): move getWhereQuery and dependencies to a different file
